forgotten and verify page restructured
