+++ src/components/React-Slick/react-slick.scss
@@ -1,5 +1,6 @@
 @import "../../colors.scss";
 @import "../../mixins.scss";
+
 .home-modal {
   width: 100vw;
   height: 100vh;
