+++ src/components/User-Account/user-account.scss
@@ -1,0 +1,114 @@
+@import "../../colors.scss";
+@import "../../mixins.scss";
+
+.picture-container {
+  @include flex(row nowrap);
+  position: fixed;
+  right: 3vw;
+  top: 3vh;
+  width: 45px;
+  height: 45px;
+  border: 2px solid $green;
+  border-radius: 50%;
+  @media (min-width: 450px) {
+    right: 3vw; //for bigger screeens
+  }
+  button {
+    position: relative;
+    width: 90%;
+    height: 90%;
+    color: white;
+    font-size: 0.75rem;
+    text-align: center;
+    font-weight: bold;
+    border: none;
+    border-radius: 50%;
+    background: $blue;
+    cursor: pointer;
+    img {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+      border-radius: 50%;
+    }
+  }
+}
+
+.user-details {
+  @include border();
+  position: fixed;
+  z-index: 99;
+  width: 60vw;
+  height: 100vh;
+  min-height: 100vh;
+  overflow: visible;
+  background: var(--white, $white);
+  opacity: 0.9;
+  font-size: 2rem;
+  padding: 0 1rem;
+  button {
+    cursor: pointer;
+    transition: transform 1s;
+  }
+  button:hover {
+    transform: scale(1.1);
+  }
+  button:active {
+    transform: scale(0.5);
+  }
+  .btn-container {
+    height: 10%;
+    width: 100%;
+    display: flex;
+    flex-flow: row nowrap;
+    align-items: center;
+  }
+  .close {
+    @include flex(row nowrap);
+    width: 50px;
+    height: 50px;
+    background: $button;
+    color: var(--black, $black);
+    font-size: 1.7rem;
+    border: none;
+    border-radius: 50%;
+  }
+
+  .update-details {
+    @include flex(column nowrap);
+    width: 90%;
+    height: 90%;
+    margin-left: calc((100% - 90%) / 2);
+    div {
+      width: 100%;
+      div {
+        height: 5px;
+        width: 30%;
+        margin-left: calc((100% - 30%) / 2);
+        border-radius: 0 0 100% 100%;
+        background: $green;
+      }
+    }
+    button {
+      font-size: 1.5rem;
+      background: transparent;
+      width: 100%;
+      border: none;
+      color: var(--black, $black);
+      a {
+        @include flex(row nowrap);
+        font-size: 1.5rem;
+        width: 100%;
+        padding: 2rem;
+        color: var(--black, $black);
+        text-decoration: none;
+      }
+    }
+    button:last-child {
+      padding: 2rem;
+    }
+  }
+}
