+++ src/pages/Update/UpdatePassword.jsx
@@ -2,9 +2,9 @@ import { useEffect, useRef } from "react";
 import { useState } from "react";
 import { useSelector } from "react-redux";
 import { Link, Navigate } from "react-router-dom";
-import animation from "../animations/popupbox";
-import LoadingScreen from "../components/loadingScreen";
-import { apiURL } from "../store/actions";
+import animation from "../../animations/popupbox";
+import LoadingScreen from "../../components/loadingScreen";
+import { apiURL } from "../../store/actions";
 
 function UpdatePassword() {
   const currentUser = useSelector((state) => state.todos.AllUserInfo);
