+++ src/styles/_toasifypop.scss
@@ -1,0 +1,29 @@
+//animated toast pop out style
+.pop-out {
+  @include border();
+  position: absolute;
+  top: 5%;
+  right: 5%;
+  width: 80vw;
+  z-index: 100;
+  background: var(--white, $white);
+  color: var(--black, $black);
+  padding: 1rem 0.3rem 0.5rem;
+  border-radius: 5px;
+  font-size: 1.3rem;
+  text-indent: 1rem;
+  .bad,
+  .good {
+    border-radius: 3px;
+    font-size: 0.3rem;
+    margin-top: 1%;
+    width: 100%;
+    height: 2%;
+    background: $red;
+    color: $red;
+  }
+  .good {
+    color: $green;
+    background: $green;
+  }
+}
